Use uint types in southbound-impl
This is a mostly-straightforward migration to use proper unsigned
types. We also fix some use of lists where it is convenient.

+++ southbound/southbound-impl/src/main/java/org/opendaylight/ovsdb/southbound/transactions/md/OvsdbQueueUpdateCommand.java
@@ -41,6 +41,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.ovsdb.re
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.Node;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
 import org.opendaylight.yangtools.yang.binding.KeyedInstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint8;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -91,7 +92,7 @@ public class OvsdbQueueUpdateCommand extends AbstractTransactionCommand {
                 queuesBuilder.setQueueUuid(new Uuid(entry.getKey().toString()));
                 Collection<Long> dscp = queue.getDscpColumn().getData();
                 if (!dscp.isEmpty()) {
-                    queuesBuilder.setDscp(dscp.iterator().next().shortValue());
+                    queuesBuilder.setDscp(Uint8.valueOf(dscp.iterator().next().shortValue()));
                 }
                 Queue oldQueue = oldQueueRows.get(entry.getKey());
                 setOtherConfig(transaction, queuesBuilder, oldQueue, queue, nodeIId);
