+++ southbound/southbound-impl/src/test/java/org/opendaylight/ovsdb/southbound/transactions/md/OvsdbPortUpdateCommandTest.java
@@ -81,6 +81,8 @@ import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.node.TerminationPointBuilder;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.node.TerminationPointKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint16;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.api.support.membermodification.MemberMatcher;
 import org.powermock.api.support.membermodification.MemberModifier;
@@ -545,10 +547,10 @@ public class OvsdbPortUpdateCommandTest {
         when(column.getData()).thenReturn(ofPorts);
         OvsdbTerminationPointAugmentationBuilder ovsdbTerminationPointBuilder = mock(
                 OvsdbTerminationPointAugmentationBuilder.class);
-        when(ovsdbTerminationPointBuilder.setOfport(any(Long.class))).thenReturn(ovsdbTerminationPointBuilder);
+        when(ovsdbTerminationPointBuilder.setOfport(any(Uint32.class))).thenReturn(ovsdbTerminationPointBuilder);
         when(interf.getName()).thenReturn(INTERFACE_NAME);
         Whitebox.invokeMethod(ovsdbPortUpdateCommand, "updateOfPort", interf, ovsdbTerminationPointBuilder);
-        verify(ovsdbTerminationPointBuilder).setOfport(any(Long.class));
+        verify(ovsdbTerminationPointBuilder).setOfport(any(Uint32.class));
     }
 
     @SuppressWarnings("unchecked")
@@ -562,11 +564,11 @@ public class OvsdbPortUpdateCommandTest {
         when(column.getData()).thenReturn(ofPortRequests);
         OvsdbTerminationPointAugmentationBuilder ovsdbTerminationPointBuilder = mock(
                 OvsdbTerminationPointAugmentationBuilder.class);
-        when(ovsdbTerminationPointBuilder.setOfportRequest(any(Integer.class)))
+        when(ovsdbTerminationPointBuilder.setOfportRequest(any(Uint16.class)))
                 .thenReturn(ovsdbTerminationPointBuilder);
         when(interf.getName()).thenReturn(INTERFACE_NAME);
         Whitebox.invokeMethod(ovsdbPortUpdateCommand, "updateOfPortRequest", interf, ovsdbTerminationPointBuilder);
-        verify(ovsdbTerminationPointBuilder).setOfportRequest(any(Integer.class));
+        verify(ovsdbTerminationPointBuilder).setOfportRequest(any(Uint16.class));
     }
 
     @SuppressWarnings("unchecked")
